requirements.txt
altgraph==0.10.2
attrs==15.2.0
bdist-mpkg==0.5.0
cssselect==0.9.1
request==2.10.0
lxml==3.6.0
macholib==1.5.1
matplotlib==1.3.1
modulegraph==0.10.4
numpy==1.8.0rc1
py2app==0.7.3
pyasn1==0.1.9
pyasn1-modules==0.0.8
pyOpenSSL==0.13.1
pyparsing==2.0.1
python-dateutil==1.5
pytz==2013.7
queuelib==1.4.2
scipy==0.13.0b1
service-identity==16.0.0
six==1.10.0
Twisted==16.0.0
vboxapi==1.0
virtualenv==15.0.1
w3lib==1.13.0
xattr==0.6.4
zope.interface==4.1.1
